clustering 代码调整
