+++ Service/Clustering/src/main/java/com/hzgc/service/clustering/bean/ClusteringVO.java
@@ -1,4 +1,4 @@
-package com.hzgc.service.clustering.dto;
+package com.hzgc.service.clustering.bean;
 
 import com.hzgc.common.service.clustering.AlarmInfo;
 import com.hzgc.service.clustering.service.ClusteringInfo;
@@ -9,7 +9,7 @@ import java.util.List;
 /**
  * 聚类查询返回前台封装
  */
-public class ClusteringDTO implements Serializable {
+public class ClusteringVO implements Serializable {
 
     /**
      * 建议迁入人口首页查询返回信息
@@ -52,7 +52,7 @@ public class ClusteringDTO implements Serializable {
 
     @Override
     public String toString() {
-        return "ClusteringDTO{" +
+        return "ClusteringVO{" +
                 "clusteringInfo=" + clusteringInfo +
                 ", alarmInfoList=" + alarmInfoList +
                 ", alarmIdList=" + alarmIdList +
